Azure migration
Previously the signup sheet was hosted on a UQCS cloud instance (uqcs1). This was difficult to manage, subject to hosting by UQ, and not as nice to maintain. Now the signup sheet is hosted as an Azure app service, and has been Dockerised (available via ghcr.io/uqcomputingsociety/signup). Documentation was also updated to reflect these changes.